nuke manifests/
yes, these are useful. but they're not kept up to date, not CI'd, and
honestly aren't worth CI'ing. the cloudformation template should be
covered by BOSH docs along with the rest of the IaaSes, and the rest of
the manifests are prone to becoming stale.

the source of truth is at:

    https://concourse.ci/clusters-with-bosh.html