bin: ensure 'gdn' is killed if the parent dies
normally, signals are trapped and forwarded along, enabling a clean
shutdown of both the parent process and 'gdn'. however, if the parent
process dies unexpectedly, we definitely don't want to leave the 'gdn'
process around, because the parent process is going to want to start its
own when it's revived.

we'll send SIGKILL to ensure the process dies - there's no point being
nice about it and sending SIGTERM because for this to be happening
something is probably already going wrong, possibly preventing the
process from shutting down in the first place. so the safest bet is to
just make sure it dies so we can at least cleanly restart.