atc: add log sink to emit error log metrics
Making use of a custom log sink we're able to capture
every log of level `error` that happens throughout the
application and increase the counter based on that.

As it might be the case that the cardinality of error logs
can get too high when paired with a very large number of
ATCs (thousands), leaving this as an opt-in feature.

concourse/concourse#2712

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>