tsa: don't drain connection when landing/retiring
otherwise builds marked 'interruptible' will never be interrupted. we
should only drain if heartbeating is explicitly interrupted, as that's
likely to be from worker rebalancing which we don't want to disrupt
normal operations

there are a *ton* of topgun changes here, a lot of it is just spring
cleaning triggered by failures i was observing locally. there is now
only one manifest, and it always configures separate VMs for
web/db/worker. this way the ops files can all be used since they can
always assume the same deployment scheme.