ci: back to windows 2016
this might not be necessary, but things were failing on 1803 and it
wasn't clear why. it was probably due to the mingw path being different,
but for whatever reason it didn't give an obvious error message. let's
just stick with 2016 for now until things are stable, to avoid changing
too many things at once.