bin/worker: adds health checker ifrit server
Adds a server process to the worker command so that
we can have an endpoint to be reached to healthchecking
the worker.

Internally, such endpoint makes a request to both
`garden` and `baggageclaim`, ensuring that both of
them are up.

It also Includes k8s deployment script for testing

concourse/concourse#2753

Signed-off-by: Topher Bullock <cbullock@pivotal.io>