ci: makes max atc waiting time be configurable for smoke tests
In the case of testing k8s upgrades, sometimes it takes more than 60 1s
ticks (1m) to have the full uprade done (with `web` having both the
liveness and readiness probes passing).

After this commit, a `smoke` task is not able to specify MAX_TICKS
(keeping the default of 60 ticks) to change such value.