atc: adds worker state metric to prometheus emitter
This PR adds a missing metric that was added before (worker state).

In the case of `prometheus`, it gets named as `concourse_worker_info`,
where `state` is just a label that we can use to aggregate on while
still having per-worker information (which can be further increased
later).

concourse/concourse#2499