atc: merge wrccs into rccs
Removed the worker_resource_config_check_sessions table from and added
the worker_base_resource_type_id column to the
resource_config_check_sessions table. These tables were able to be
merged into one because of the removal of the team_id column on the
wrccs table. The only reason why we had the separation of the two tables
was because there could be one rccs with many wrccs (when there are
multiple teams using that one rccs so we need to make multiple check
containers for each team). Now that check containers are global, there
isn't the need for the two separate tables. This change came about from
a bug that happened randomly at times when the rccs would expire before
the wrccs gets created (which would error b/c wrccs has a foreign key
constraint on rccs id).

fixes #1598