atc: clearer error for no global workers available
The error that would've been shown before this change would've been
that there are no workers available but if the user had a team worker
but no global workers, they would be confused on why concourse was
saying they did not have any workers available. This commit makes it
clearer that the problem is that they do not have any global workers
available for check containers to be placed on. Also changed the general
no workers available errors to not include the available workers because
that might be a security risk of listing the workers to users not within
the team.