ci: adds `--recreate-pods` flag to k8s smoke deploy
Previously, if a deploy different from another just in terms of changing
the contents of secrets, no pods would be recreated.

By making use of `--recreate-pods` we're able to do the same (in a
non highly-available manner) as setting an annotation in the deployment
that would contain a `sha256sum` of the contents of the secrets.

Eventually we might go with that approach (we used to do in our fork)
soon - need to prepare a PR first.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>