topgun: makes use of default helm args in k8s
Previously, although we were constructing a list that contained default
arguments to be used by the helm deployment of concourse, we forgot to
pass them to the command execution (thus, not using them).

This commit fixes that and also adds some other arguments to reduce the
delays that the default probes present.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>