atc: fix resourceconfig foreign key constraint bug
Within the put step and radar, we used to find or create a resource
config and then further down in the codepath, we would use that resource
config to fill in the resource_config_id column in either the resource
config check sessions table or the resource config versions table. In
the time between the creation of the resource config and the filling in
the column, the resource config could have become garbage collected
resulting in the query to fill in the resource_config_id column failing
because of the foreign key constraint. This commit fixes this by
creating the resource config and filling the column in within the same
transaction.

We also added an ON DELETE RESTRICT constraint on the resource config id
in the resource types table because we should not be garbage collecting
resource configs that are being referenced by a resource type.