topgun: adds baggageclaim driver tests for k8s suite
This commit adds explicit tests for the three different baggageclaim
drivers that we support: naive, overlay, and btrfs.

Right now, btrfs seems to be an issue for those running under GKE with
the container-optimized image type given that they don't ship it with
the btrfs module activated (a requirement for us to test such module) -
thus, the commit leaves `btrfs` skipped for now.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>