topgun & ci: adds bbr tests
This commit adds two tests for testing the BBR configuration that
Concourse has.

It introduces two tests:

1. a test that verifies whether backing up and restoring works when
starting from a clean state and restoring back to that same state; and

2. another that checks if we can go back to a clean state after having a
pipeline set.

The first test consistently passes, while the second fails - to be fixed
in upcoming commits.

note.: we added an `ops` file that makes the postgres version static so
BBR doesn't fail when parsing the version of the latest postgres (see
[1]).

[1]: https://github.com/cloudfoundry-incubator/backup-and-restore-sdk-release/issues/14

concourse/concourse#2332

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bishoy youssef <byoussef@pivotal.io>