atc: remove custom type scan in resource scanner
Instead of manually running a resource type scan within the resource
scanner when the resource uses a custom type with not versions, it will
wait on the resource type to recieve a version from it's own resource
type scanning interval. This fixes the problem when many resources use
one resource type that has no versions yet, it will start off many scans
which all try to acquire the same lock. Resource type scanning also now
differs from the resource scanner because it does not loop if it fails
to acquire the lock, it will just exit with the error. We are able to do
this because even if global resources is disabled, resource type
versions are still shared within a concourse deployment so there is no
real need for them to pull in their own versions for each pipeline
resource.

[#3241]

Signed-off-by: Clara Fu <cfu@pivotal.io>