atc: fly exec builds get run by the scheduler
instead of a calling enginebuild.Resume in the handler

*wip note*

we are currently hardcoding the engine as 'exec.v2' when creating
started builds in the databse. We should probably leave this up to
the dbengine to update the engine field if its empty.

we should also probably change the engine_metadata to private_plan
and reintroduce a metadata field later if needed.

concourse/concourse#3240

Signed-off-by: Josh Winters <jwinters@pivotal.io>
Co-authored-by: Mark Huang <mhuang@pivotal.io>