atc: remove worker/pool client
The worker.Client interface only existed because there were parts of the
code that wanted to choose a worker and then find or create a container
but there were other parts that wanted to use a specific worker and then
find or create container on that worker. There were these two paths but
in order to use the same code, both the pool and worker objects were
adjusted in order to fit this worker Client interface. This commit
removes this dependency by having the pool object find or choose a
worker and then on the worker object find or create a container on that
chosen worker.

Also, get containers are now being placed with the chosen placement
strategy instead of always randomly placing them.

Signed-off-by: Clara Fu <cfu@pivotal.io>