refactor worker rebalancing to fix draining
previously, if a worker timed out draining and came back while still in
landing/retiring state, it would remain in that state forever, so that
rebalancing wouldn't accidentally switch it to 'running'

instead, this commit prevents rebalancing during draining. this is done
by merging the drain runner into the beacon and synchronizing so that
draining can't happen with a concurrent rebalance registration
in-flight.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>