ci: k8s-smoke - use naive BC driver and make liveness probes fail faster
By using the `naive` bc driver we can be sure that regardless of which
node pool ends up running the worker pod, it'll work. This is necessary
if we're not setting specific node selectors. As we don't run any big
workloads on this deployment, perf wouldn't be a problem.

Reducing `web.livenessProbe` we're able to leverage the restarting of
`concourse web` more often if it takes long for `postgresql` to get
started (so we don't get impacted by the exponential backoffs).

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>