rename --tsa-peer-ip to --tsa-peer-address
this can also be a DNS address, not just an IP

this looks backwards-incompatible but in practice this flag was always
auto-set by taking the host from `--peer-url`. that flag has now been
removed, and --tsa-peer-address is to be used instead, so we might as
well get the name right.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>