ci: add release pipeline template
this was successfully used for 5.0.1.

release pipelines are responsible for issuing patch versions for
supported MAJOR.MINOR version lines.

some learnings from this process:

* the main pipeline should no longer be used for patch releases - it
  should only be used for the initial MAJOR.MINOR.0 release, and after
  that point we should have a release pipeline for subsequent patches.

* syncing CI config (tasks, pipelines, etc.) across release branches is
  a huge pain. in this case, changes were made to the release branch ci/
  stuff to make them more flexible for use in release branches, but we
  obviously want these changes on `master`, too.

  maybe these changes should have been made on `master` first and then
  cherry-picked to the release branches? that's kind of easy to say in
  hindsight - it's hard to get a feedback loop without "doing it live"
  on the release branch.

one alternative to syncing CI config would be to treat each release
branch as a 'snapshot' of the minor release's point in time, and
maintain the CI config as it was. this is how we've treated hotfixes in
the past, but it feels a bit more like 'maintaining' minor releases vs.
'supporting' them, and in this case there were quite a few fixes made to
ci/ that we wanted for 5.0.1. maybe this is just the complicated reality
of supporting multiple releases: sometimes you want to share config,
sometimes you don't.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>