ci: don't bump cbd in release pipelines
instead, split the main pipeline 'publish-bosh-deployment' job in half:

* 'promote-cbd' which promotes 'develop' to 'master, which should run
  after we ship every time.
* 'bump-cbd-versions' which just bumps versions.yml *on master* based on
  the latest version available on bosh.io, which we can let run if a
  release pipeline bumped a minor version, too

theoretically we might want a 'promote-cbd' for release pipelines, but
only if the release pipeline is for the latest minor release. that's
hard to automate, so for now maybe we should just manually promote it
when necessary as it's not often necessary for patch releases anyway.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>