atc: actually implement semver-y event parsing
up until now we've been using MAJOR.MINOR versions for events, but it's
been pretty much useless as version lookup was by direct string
equality.

this changes the event parsing to be compatible as long as the major
version is the same, and removes 'deprecated' events for older minor
versions, since they should no longer be necessary.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>