atc: add multi version constraining on builds
In the algorithm, we used to constrain only on one of the candidate
versions in order to find all the builds of a passed job that also
has a build that has outputted that version. We can possibly further
improve on this to make it not waste time on grabbing builds that match
on the current version candidate but will fail on other version
candidates. To that point, we made it so we will only grab builds that
have outputted the same versions for all the candidates that depend on
that passed job.

[#3602]

Co-authored-by: Alex Suraci <suraci.alex@gmail.com>