atc: fix serial groups bug
If a build is set to scheduled but not yet started, the max in flighter
will see that scheduled build and will determine that the max in flight
is reached. This commit fixes that by checking if the build is scheduled
already, if it is then it skips over all the max in flight checking.
Also removed the max in flight package and put everything within the
ScheduleBuild method inside the job object, that way we will check the
max in flight and schedule the build all in one transaction.

[#3602, #4065]

Signed-off-by: Clara Fu <cfu@pivotal.io>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>