topgun/k8s: move random seed to release+ns setup
Previously, we had the seed for random being configured only at the
suite level, which lead to us having problems with name colisions.

Now, having it being seeded every time (+ having a bigger random number
size) we need a random name, we're able to better ensure that we won't
have collisions even when running in parallel.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>
Co-authored-by: Bishoy Youssef <byoussef@pivotal.io>